Issue #12: Specify tablePrefix on all spring beans
Pass the table prefix to all Spring beans that require
it to be set when trying to change it. Additionally require
the configuration of the 'database' config option and when
not specified, fall back to h2.

This is now required so that it can be passed to the
jobRepository configuration.

Also fix SQL commit warning when loading tables by
executing script within a transaction.